pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trufnetwork-adapters"
version = "0.1.0"
description = "A collection of prefect tasks and examples ingesting data into TSN"
readme = "README.md"
dependencies = [
    "prefect-client>=3.3.1",
    "truflation-data@git+https://github.com/truflation/truflation.git@deploy-production-20241008",
    "trufnetwork-sdk-py@git+https://github.com/trufnetwork/sdk-py.git@bdc4d187add38228b8051ef7faade0d19ee974be",
    "PyGithub",
    "pandera>=0.23.1",
    "pandas",
    "prefect-sqlalchemy",
    "SQLAlchemy",
    "psycopg2-binary",
]

[project.optional-dependencies]
argentina = [
    "beautifulsoup4",
    "dateparser",
    "tqdm",
    "boto3",
    "prefect-aws>=0.5.3",
]
dev = [
    # we need this for the testing harness
    "prefect>=3.3.1",
    "pytest-timeout",
    "pytest-mock",
    "black",
    "isort",
    "ruff",
    "moto",
    "pytest-postgresql",
    "docker",
]

[tool.setuptools]
packages = {find = {where = ["src"]}}
package-dir = {"" = "."}

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: mark test as integration test",
    "slow: mark test as slow to run",
]
pythonpath = ["src"]
addopts = "--import-mode=importlib -v -s"
testpaths = ["tests"]

[tool.ruff]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "RUF",  # ruff-specific rules
    "TID",  # flake8-tidy-imports
]
line-length = 120
target-version = "py39"
src = ["src"]

# Ignore `__init__.py` files
exclude = ["__init__.py"]

[tool.ruff.isort]
known-first-party = ["tsn_adapters"]
force-sort-within-sections = true
combine-as-imports = true

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
